+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- CSS classes for reaching a sku variation name and value.
+
 ## [3.28.1] - 2019-04-25
 ### Fixed
 - Fix propType `ShippingSimulator` error, removed warning from console.